Add unit tests for NSColor extensions #257 fix
Added tests for NSColor, renamed to ColorExtensionsTests
